% Generated by generate-rd-files.py
\name{build_symbol_index}
\alias{build_symbol_index}
\title{Build Symbol Index from Cached Data}
\description{
Build Symbol Index from Cached Data
}
\usage{
build_symbol_index(hgnc_data = NULL)
}
\arguments{
\item{hgnc_data}{(parameter description)}
}
\details{
This is an automatically generated documentation file.
For complete documentation, run: roxygen2::roxygenise()
}
