% Generated by generate-rd-files.py
\name{clear_hgnc_cache}
\alias{clear_hgnc_cache}
\title{Rate Limiter for HGNC API Requests}
\description{
Rate Limiter for HGNC API Requests
}
\usage{
clear_hgnc_cache()
}
\details{
This is an automatically generated documentation file.
For complete documentation, run: roxygen2::roxygenise()
}
