% Generated by generate-rd-files.py
\name{download_hgnc_data}
\alias{download_hgnc_data}
\title{Get HGNC cache directory}
\description{
Get HGNC cache directory
}
\usage{
download_hgnc_data(url = .HGNC_DATA_URL, force = FALSE)
}
\arguments{
\item{url}{(parameter description)}
\item{force}{(parameter description)}
}
\details{
This is an automatically generated documentation file.
For complete documentation, run: roxygen2::roxygenise()
}
