% Generated by generate-rd-files.py
\name{get_hgnc_cache_dir}
\alias{get_hgnc_cache_dir}
\title{Get HGNC cache directory}
\description{
Get HGNC cache directory
}
\usage{
get_hgnc_cache_dir()
}
\details{
This is an automatically generated documentation file.
For complete documentation, run: roxygen2::roxygenise()
}
