% Generated by generate-rd-files.py
\name{get_hgnc_cache_info}
\alias{get_hgnc_cache_info}
\title{Get HGNC cache directory}
\description{
Get HGNC cache directory
}
\usage{
get_hgnc_cache_info()
}
\details{
This is an automatically generated documentation file.
For complete documentation, run: roxygen2::roxygenise()
}
