% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgnc_data.R
\name{get_hgnc_cache_path}
\alias{get_hgnc_cache_path}
\title{Get path to cached HGNC data file}
\usage{
get_hgnc_cache_path()
}
\value{
Character string with the path to the cached data file
}
\description{
Get path to cached HGNC data file
}
\keyword{internal}
