% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgnc_data.R
\name{get_hgnc_metadata_path}
\alias{get_hgnc_metadata_path}
\title{Get path to cache metadata file}
\usage{
get_hgnc_metadata_path()
}
\value{
Character string with the path to the metadata file
}
\description{
Get path to cache metadata file
}
\keyword{internal}
