% Generated by generate-rd-files.py
\name{hgnc_changes}
\alias{hgnc_changes}
\title{Track Gene Nomenclature Changes}
\description{
Track Gene Nomenclature Changes
}
\usage{
hgnc_changes(since, fields = c("symbol", "name", "status"), change_type = c("all", "symbol", "name", "status", "modified"), use_cache = TRUE)
}
\arguments{
\item{since}{(parameter description)}
\item{fields}{(parameter description)}
\item{change_type}{(parameter description)}
\item{use_cache}{(parameter description)}
}
\details{
This is an automatically generated documentation file.
For complete documentation, run: roxygen2::roxygenise()
}
