% Generated by generate-rd-files.py
\name{hgnc_fetch}
\alias{hgnc_fetch}
\title{Search for Genes in HGNC Database}
\description{
Search for Genes in HGNC Database
}
\usage{
hgnc_fetch(field, term)
}
\arguments{
\item{field}{(parameter description)}
\item{term}{(parameter description)}
}
\details{
This is an automatically generated documentation file.
For complete documentation, run: roxygen2::roxygenise()
}
