% Generated by generate-rd-files.py
\name{hgnc_find}
\alias{hgnc_find}
\title{Search for Genes in HGNC Database}
\description{
Search for Genes in HGNC Database
}
\usage{
hgnc_find(query, filters = NULL, limit = 100)
}
\arguments{
\item{query}{(parameter description)}
\item{filters}{(parameter description)}
\item{limit}{(parameter description)}
}
\details{
This is an automatically generated documentation file.
For complete documentation, run: roxygen2::roxygenise()
}
