% Generated by generate-rd-files.py
\name{hgnc_get_changes_summary}
\alias{hgnc_get_changes_summary}
\title{HGNC MCP Resource Helpers}
\description{
HGNC MCP Resource Helpers
}
\usage{
hgnc_get_changes_summary(since, format = c("json", "markdown", "text"), change_type = "all", max_results = 100)
}
\arguments{
\item{since}{(parameter description)}
\item{format}{(parameter description)}
\item{change_type}{(parameter description)}
\item{max_results}{(parameter description)}
}
\details{
This is an automatically generated documentation file.
For complete documentation, run: roxygen2::roxygenise()
}
