% Generated by generate-rd-files.py
\name{hgnc_get_gene_card}
\alias{hgnc_get_gene_card}
\title{HGNC MCP Resource Helpers}
\description{
HGNC MCP Resource Helpers
}
\usage{
hgnc_get_gene_card(hgnc_id, format = c("json", "markdown", "text"))
}
\arguments{
\item{hgnc_id}{(parameter description)}
\item{format}{(parameter description)}
}
\details{
This is an automatically generated documentation file.
For complete documentation, run: roxygen2::roxygenise()
}
