% Generated by generate-rd-files.py
\name{hgnc_resolve_symbol}
\alias{hgnc_resolve_symbol}
\title{Search for Genes in HGNC Database}
\description{
Search for Genes in HGNC Database
}
\usage{
hgnc_resolve_symbol(symbol, mode = "lenient", return_record = FALSE)
}
\arguments{
\item{symbol}{(parameter description)}
\item{mode}{(parameter description)}
\item{return_record}{(parameter description)}
}
\details{
This is an automatically generated documentation file.
For complete documentation, run: roxygen2::roxygenise()
}
