% Generated by generate-rd-files.py
\name{hgnc_rest_info}
\alias{hgnc_rest_info}
\title{Get HGNC REST API Information (Cached)}
\description{
Cached version of hgnc_rest_info_uncached(). Uses memoise to cache results.
}
\usage{
hgnc_rest_info()
}
\details{
This is a memoised wrapper around hgnc_rest_info_uncached().
For complete documentation, see hgnc_rest_info_uncached() or run: roxygen2::roxygenise()
}
\seealso{
\code{\link{hgnc_rest_info_uncached}}
}
