% Generated by generate-rd-files.py
\name{hgnc_rest_info_uncached}
\alias{hgnc_rest_info_uncached}
\title{Rate Limiter for HGNC API Requests}
\description{
Rate Limiter for HGNC API Requests
}
\usage{
hgnc_rest_info_uncached()
}
\details{
This is an automatically generated documentation file.
For complete documentation, run: roxygen2::roxygenise()
}
