% Generated by generate-rd-files.py
\name{hgnc_search_groups}
\alias{hgnc_search_groups}
\title{Get Members of a Gene Group}
\description{
Get Members of a Gene Group
}
\usage{
hgnc_search_groups(query, limit = 100)
}
\arguments{
\item{query}{(parameter description)}
\item{limit}{(parameter description)}
}
\details{
This is an automatically generated documentation file.
For complete documentation, run: roxygen2::roxygenise()
}
