% Generated by generate-rd-files.py
\name{hgnc_validate_panel}
\alias{hgnc_validate_panel}
\title{Track Gene Nomenclature Changes}
\description{
Track Gene Nomenclature Changes
}
\usage{
hgnc_validate_panel(items, policy = "HGNC", suggest_replacements = TRUE, include_dates = TRUE, index = NULL)
}
\arguments{
\item{items}{(parameter description)}
\item{policy}{(parameter description)}
\item{suggest_replacements}{(parameter description)}
\item{include_dates}{(parameter description)}
\item{index}{(parameter description)}
}
\details{
This is an automatically generated documentation file.
For complete documentation, run: roxygen2::roxygenise()
}
