% Generated by generate-rd-files.py
\name{hgnc_xrefs}
\alias{hgnc_xrefs}
\title{Search for Genes in HGNC Database}
\description{
Search for Genes in HGNC Database
}
\usage{
hgnc_xrefs(id_or_symbol)
}
\arguments{
\item{id_or_symbol}{(parameter description)}
}
\details{
This is an automatically generated documentation file.
For complete documentation, run: roxygen2::roxygenise()
}
