% Generated by generate-rd-files.py
\name{is_hgnc_cache_fresh}
\alias{is_hgnc_cache_fresh}
\title{Get HGNC cache directory}
\description{
Get HGNC cache directory
}
\usage{
is_hgnc_cache_fresh(max_age_days = 30)
}
\arguments{
\item{max_age_days}{(parameter description)}
}
\details{
This is an automatically generated documentation file.
For complete documentation, run: roxygen2::roxygenise()
}
