% Generated by generate-rd-files.py
\name{load_hgnc_data}
\alias{load_hgnc_data}
\title{Get HGNC cache directory}
\description{
Get HGNC cache directory
}
\usage{
load_hgnc_data(max_age_days = 30, force = FALSE)
}
\arguments{
\item{max_age_days}{(parameter description)}
\item{force}{(parameter description)}
}
\details{
This is an automatically generated documentation file.
For complete documentation, run: roxygen2::roxygenise()
}
