% Generated by generate-rd-files.py
\name{start_hgnc_mcp_server}
\alias{start_hgnc_mcp_server}
\title{Start HGNC MCP Server}
\description{
Start HGNC MCP Server
}
\usage{
start_hgnc_mcp_server(port = 8080, host = "0.0.0.0", swagger = TRUE, quiet = FALSE, ...)
}
\arguments{
\item{port}{(parameter description)}
\item{host}{(parameter description)}
\item{swagger}{(parameter description)}
\item{quiet}{(parameter description)}
\item{...}{(parameter description)}
}
\details{
This is an automatically generated documentation file.
For complete documentation, run: roxygen2::roxygenise()
}
